+++ examples/shors/Cargo.toml
@@ -8,3 +8,10 @@ quasim = { path = "../../quasim" }
 gcd = "2.3.0"
 rand = "0.10.0"
 num-integer = "0.1.46"
+
+[dev-dependencies]
+divan = "0.1.21"
+
+[[bench]]
+name = "shors-benchmark"
+harness = false
